seperate defs of Robot from World; get rid of World::robot_list_
