+++ xrobot/render_engine/visualization.cpp
@@ -115,8 +115,9 @@ void Visualization::DrawPointCloud(const GLuint tex, Camera* camera,
 }
 
 void Visualization::DrawRootAABB(RenderWorld* world, const Shader& shader) {
-	for (size_t i = 0; i < world->size(); i++) {
-		RenderBody* body = world->render_body_ptr(i);
+    world->robot_iteration_begin();
+    while (world->has_next_robot()) {
+        RenderBody* body = world->next_robot();
 		RenderPart * part = body->render_root_ptr();
 		if (part && !body->is_recycled()) {
 			glm::vec3 aabb_min, aabb_max;
@@ -304,12 +305,12 @@ void Visualization::Draw(RenderWorld* world, const Shader& shader) {
 		}
 	};
 
-	for (size_t i = 0; i < world->size(); ++i) {
-		RenderBody* body = world->render_body_ptr(i);
-
-		if(body->is_hiding())
+   world->robot_iteration_begin();
+    while (world->has_next_robot()) {
+        RenderBody* body = world->next_robot();
+		if(body->is_hiding()) {
 			continue;
-
+        }
 		// Root
 		RenderPart* root = body->render_root_ptr();
 		if (root && !body->is_recycled()) {
